requirements.txt
Flask==3.0.0
Flask-SQLAlchemy==3.1.1
Flask-Mail==0.9.1
python-dotenv==1.0.0
firebase-admin==6.4.0
pyrebase4==4.7.1
xendit==2.9.0
requests==2.31.0
